Split buffer
